// react-burger/src/utils/types/ingredient-types.tsx
type TIngredientType = "bun" | "sauce" | "main";

export type TIngredient = {
  readonly _id: string;
  readonly name: string;
  readonly type: TIngredientType;
  readonly proteins: number;
  readonly fat: number;
  readonly carbohydrates: number;
  readonly calories: number;
  readonly price: number;
  readonly image: string;
  readonly image_mobile: string;
  readonly image_large: string;
  readonly __v: number;
};

export type TIngredientWithID = TIngredient & { id: string };